done complaint

+++ src/main/java/org/example/ecommerce/entity/Promotion.java
@@ -4,7 +4,6 @@ import jakarta.persistence.*;
 import lombok.Data;
 import org.hibernate.annotations.Nationalized;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 import java.util.List;
@@ -50,7 +49,7 @@ public class Promotion {
     @Column(name = "per_user_limit")
     private Integer perUserLimit;
     private Double value;
-    private BigDecimal revenue;
+    private Double revenue;
     private String status;
     private Integer orders;
     @ManyToMany(cascade = {CascadeType.DETACH, CascadeType.MERGE, CascadeType.PERSIST, CascadeType.REFRESH})
